# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Scanbo FHIR Converter API
  version: 1.1.0
  description: |
    Convert custom JSON, HL7v2, CDA, CSV, and Excel into FHIR R5 Bundles.
servers:
  - url: http://localhost:3000
    description: Local
...
paths:
  /json:
    post:
      tags: [Custom JSON]
      summary: Convert custom JSON to FHIR R5 Bundle
      description: |
        Accepts global custom JSON (snake_case). Legacy camelCase payloads are still supported for backward compatibility.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CustomJsonPatientOnly'
                - $ref: '#/components/schemas/CustomJsonEncounterOnly'
                - $ref: '#/components/schemas/CustomJsonMedicationOnly'
                - $ref: '#/components/schemas/CustomJsonMedicationRequestOnly'
                - $ref: '#/components/schemas/CustomJsonPractitionerOnly'
                - $ref: '#/components/schemas/CustomJsonPractitionerRoleOnly'
                - $ref: '#/components/schemas/CustomJsonOrganizationOnly'
                - $ref: '#/components/schemas/CustomJsonFull'
            examples:
              patient:
                summary: Patient
                value:
                  patient:
                    patient_id: P-10001
                    ihi: "1234098753"
                    name:
                      first_name: John
                      middle_name: A.
                      last_name: Doe
                    date_of_birth: 1980-05-15
                    gender: male
                    contact_info:
                      phone: "+1-555-555-5555"
                      email: john.doe@example.com
                      address:
                        street: 123 Main St
                        city: Springfield
                        state: IL
                        postal_code: 62701
                        country: USA
                    marital_status: married
                    language: English
                    ethnicity: Caucasian
                    emergency_contact:
                      name: Jane Doe
                      relationship: spouse
                      phone: "+1-555-555-5556"
                    insurance:
                      provider: Good Health Insurance
                      policy_number: GH123456789
                      coverage_start_date: 2020-01-01
                      coverage_end_date: 2025-01-01
                    medical_history:
                      allergies:
                        - Penicillin
                        - Peanuts
                      current_medications:
                        - Lisinopril
                        - Atorvastatin
                      past_illnesses:
                        - Hypertension
                        - Asthma
                      surgeries:
                        - surgery_name: Appendectomy
                          surgery_date: 2000-06-15
                        - surgery_name: Knee Replacement
                          surgery_date: 2015-08-20
              encounter:
                summary: Encounter
                value:
                  encounter:
                    encounter_id: ENC345678
                    patient_id: 12345
                    practitioner_id: 98765
                    encounter_type: Outpatient
                    reason_for_visit: Routine check-up
                    start_date: 2024-08-15T09:00:00Z
                    end_date: 2024-08-15T09:30:00Z
                    location:
                      facility_name: Healthy Heart Clinic
                      room: Room 101
                      address:
                        street: 789 Clinic Blvd
                        city: Healthyville
                        state: CA
                        postal_code: 90211
                        country: USA
                    status: completed
                    diagnoses:
                      - condition_id: COND789012
                        condition_name: Hypertension
                    services_provided:
                      - Blood pressure measurement
                      - Physical examination
                    billing_info:
                      insurance_provider: Good Health Insurance
                      policy_number: GH123456789
                      total_cost: 150
                      patient_responsibility: 30
                    follow_up:
                      recommended: true
                      next_appointment_date: 2024-09-15T09:00:00Z
                    note: Patient advised to continue current medication and follow-up in one month.
              medication:
                summary: Medication
                value:
                  medication:
                    medication_id: MED987654
                    name: Atorvastatin
                    brand_name: Lipitor
                    form: Tablet
                    strength: 20 mg
                    manufacturer:
                      name: Pfizer Inc.
                      address:
                        street: 235 East 42nd Street
                        city: New York
                        state: NY
                        postal_code: 10017
                        country: USA
                    ingredients:
                      - name: Atorvastatin Calcium
                        strength: 20 mg
                      - name: Lactose Monohydrate
                        strength: Inactive Ingredient
                    package:
                      description: Bottle of 30 tablets
                      quantity: 30
                    status: active
                    expiration_date: 2025-12-31
                    lot_number: AB123456
              medication_request:
                summary: MedicationRequest
                value:
                  medication_request:
                    medication_request_id: MRQ123456
                    patient_id: 12345
                    practitioner_id: 98765
                    medication:
                      medication_id: MED987654
                      name: Atorvastatin
                      strength: 20 mg
                      form: Tablet
                    dosage_instruction:
                      dose: 20 mg
                      route: Oral
                      frequency: Once daily
                      duration: 30 days
                    dispense_request:
                      quantity: 30
                      refills: 2
                      expected_supply_duration: 30 days
                      dispense_as_written: false
                    substitution_allowed: true
                    reason_for_prescription: Hyperlipidemia
                    status: active
                    authored_on: 2024-08-01
                    note: Patient advised to take medication in the evening.
              practitioner:
                summary: Practitioner
                value:
                  practitioner:
                    practitioner_id: 98765
                    name:
                      first_name: Emily
                      middle_name: R.
                      last_name: Smith
                    date_of_birth: 1975-10-12
                    gender: female
                    contact_info:
                      phone: "+1-555-555-7777"
                      email: emily.smith@healthcare.com
                      address:
                        street: 456 Health St
                        city: Healthyville
                        state: CA
                        postal_code: 90210
                        country: USA
                    license:
                      license_number: MD123456
                      state_issued: CA
                      issue_date: 2000-05-20
                      expiry_date: 2025-05-20
                    specialization: Cardiology
                    years_of_experience: 20
                    languages_spoken:
                      - English
                      - Spanish
                    practice_location:
                      name: Healthy Heart Clinic
                      address:
                        street: 789 Clinic Blvd
                        city: Healthyville
                        state: CA
                        postal_code: 90211
                        country: USA
                      phone: "+1-555-555-8888"
                    qualifications:
                      - degree: MD
                        institution: Harvard Medical School
                        year: 2000
                      - degree: Fellowship in Cardiology
                        institution: Johns Hopkins Hospital
                        year: 2004
              practitioner_role:
                summary: PractitionerRole
                value:
                  practitioner_role:
                    practitioner_role_id: PR123456
                    practitioner_id: 98765
                    organization_id: ORG001
                    role: Cardiologist
                    specialty: Cardiology
                    location:
                      location_id: LOC123456
                      name: Healthy Heart Clinic
                      address:
                        street: 789 Clinic Blvd
                        city: Healthyville
                        state: CA
                        postal_code: 90211
                        country: USA
                    telecom:
                      phone: "+1-555-555-7777"
                      email: emily.smith@healthyheartclinic.com
                    available_hours:
                      weekdays: 8:00 AM - 5:00 PM
                      weekends: 9:00 AM - 2:00 PM
                    service_period:
                      start_date: 2020-01-01
                      end_date: null
              organization:
                summary: Organization
                value:
                  organization:
                    organization_id: ORG001
                    name: Healthy Heart Clinic
                    type: Healthcare Provider
                    contact_info:
                      phone: "+1-555-555-8888"
                      email: contact@healthyheartclinic.com
                      address:
                        street: 789 Clinic Blvd
                        city: Healthyville
                        state: CA
                        postal_code: 90211
                        country: USA
                      fax: "+1-555-555-9999"
                      website: https://www.healthyheartclinic.com
                    departments:
                      - name: Cardiology
                        contact_phone: "+1-555-555-1234"
                      - name: Radiology
                        contact_phone: "+1-555-555-5678"
                    affiliations:
                      - name: American Heart Association
                        type: Professional Association
                      - name: National Institutes of Health
                        type: Research Partner
                    services_offered:
                      - Cardiac Consultation
                      - Echocardiography
                      - Stress Testing
                      - Cardiac Surgery
                    operating_hours:
                      weekdays: 8:00 AM - 5:00 PM
                      weekends: 9:00 AM - 2:00 PM
              fullBundle:
                summary: Full bundle
                value:
                  operation: create
                  patient:
                    patient_id: P-10001
                    name:
                      first_name: John
                      last_name: Doe
                    date_of_birth: 1980-05-15
                  encounter:
                    encounter_id: ENC345678
                    encounter_type: Outpatient
                    start_date: 2024-08-15T09:00:00Z
                    status: completed
                  medication:
                    medication_id: MED987654
                    name: Atorvastatin
                    strength: 20 mg
                  medication_request:
                    medication_request_id: MRQ123456
                    medication:
                      medication_id: MED987654
                      name: Atorvastatin
                  practitioner:
                    practitioner_id: 98765
                    name:
                      first_name: Emily
                      last_name: Smith
                  practitioner_role:
                    practitioner_role_id: PR123456
                    practitioner_id: 98765
                    role: Cardiologist
                  organization:
                    organization_id: ORG001
                    name: Healthy Heart Clinic
      responses:
        '200':
          description: FHIR R5 Bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FhirBundle'
        '400':
          description: Validation or conversion error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  summary: Schema validation error
                  value:
                    error: "JSON validation failed: <field>: <message>"
                emptyPayload:
                  summary: Empty payload error
                  value:
                    error: "JSON validation failed: At least one resource section is required (patient, encounter, medication, medication_request, practitioner, practitioner_role, organization)."
  /convert:
    post:
      tags: [Convert]
      summary: Convert mixed input (auto-detect)
      description: |
        Accepts JSON body with an input string, or raw text/XML body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertRequest'
          text/plain:
            schema:
              type: string
          application/xml:
            schema:
              type: string
      parameters:
        - in: query
          name: format
          schema:
            type: string
          description: Optional input format override
        - in: query
          name: fhirVersion
          schema:
            type: string
            enum: [r5]
          description: FHIR output version (default r5)
      responses:
        '200':
          description: FHIR R5 Bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FhirBundle'
        '400':
          description: Validation or conversion error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /convert/upload:
    post:
      tags: [Upload]
      summary: Upload file for conversion
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                format:
                  type: string
                  description: Optional file format override
                fhirVersion:
                  type: string
                  enum: [r5]
      responses:
        '200':
          description: FHIR R5 Bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FhirBundle'
        '400':
          description: Validation or conversion error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /convert/hl7:
    post:
      tags: [Convert]
      summary: Convert HL7 input to FHIR R5 Bundle
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertRequest'
      parameters:
        - in: query
          name: format
          schema:
            type: string
          description: Optional input format override
        - in: query
          name: fhirVersion
          schema:
            type: string
            enum: [r5]
          description: FHIR output version (default r5)
      responses:
        '200':
          description: FHIR R5 Bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FhirBundle'
        '400':
          description: Validation or conversion error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    ConvertRequest:
      type: object
      properties:
        input:
          type: string
        format:
          type: string
        fhirVersion:
          type: string
          enum: [r5]
      required: [input]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    CustomJsonInput:
      type: object
      description: At least one resource section is required (patient, encounter, medication, medication_request, practitioner, practitioner_role, organization).
      properties:
        operation:
          type: string
          enum: [create, update, delete]
        messageType:
          type: string
        patient:
          oneOf:
            - $ref: '#/components/schemas/CustomPatient'
            - type: array
              items:
                $ref: '#/components/schemas/CustomPatient'
        encounter:
          oneOf:
            - $ref: '#/components/schemas/CustomEncounter'
            - type: array
              items:
                $ref: '#/components/schemas/CustomEncounter'
        medication:
          oneOf:
            - $ref: '#/components/schemas/CustomMedication'
            - type: array
              items:
                $ref: '#/components/schemas/CustomMedication'
        medication_request:
          oneOf:
            - $ref: '#/components/schemas/CustomMedicationRequest'
            - type: array
              items:
                $ref: '#/components/schemas/CustomMedicationRequest'
        practitioner:
          oneOf:
            - $ref: '#/components/schemas/CustomPractitioner'
            - type: array
              items:
                $ref: '#/components/schemas/CustomPractitioner'
        practitioner_role:
          oneOf:
            - $ref: '#/components/schemas/CustomPractitionerRole'
            - type: array
              items:
                $ref: '#/components/schemas/CustomPractitionerRole'
        organization:
          oneOf:
            - $ref: '#/components/schemas/CustomOrganization'
            - type: array
              items:
                $ref: '#/components/schemas/CustomOrganization'
...
    CustomPatient:
      type: object
      properties:
        patient_id:
          type: string
        ihi:
          type: string
        name:
          $ref: '#/components/schemas/CustomPatientName'
        date_of_birth:
          type: string
          format: date
        gender:
          type: string
        contact_info:
          $ref: '#/components/schemas/CustomPatientContactInfo'
        marital_status:
          type: string
        language:
          type: string
        ethnicity:
          type: string
        emergency_contact:
          $ref: '#/components/schemas/CustomPatientEmergencyContact'
        insurance:
          $ref: '#/components/schemas/CustomPatientInsurance'
        medical_history:
          $ref: '#/components/schemas/CustomPatientMedicalHistory'
    CustomPatientName:
      type: object
      properties:
        first_name:
          type: string
        middle_name:
          type: string
        last_name:
          type: string
    CustomPatientContactInfo:
      type: object
      properties:
        phone:
          type: string
        email:
          type: string
        address:
          $ref: '#/components/schemas/CustomAddress'
    CustomPatientEmergencyContact:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
    CustomPatientInsurance:
      type: object
      properties:
        provider:
          type: string
        policy_number:
          type: string
        coverage_start_date:
          type: string
          format: date
        coverage_end_date:
          type: string
          format: date
    CustomPatientMedicalHistory:
      type: object
      properties:
        allergies:
          type: array
          items:
            type: string
        current_medications:
          type: array
          items:
            type: string
        past_illnesses:
          type: array
          items:
            type: string
        surgeries:
          type: array
          items:
            $ref: '#/components/schemas/CustomPatientSurgery'
...
    CustomEncounter:
      type: object
      properties:
        encounter_id:
          type: string
        patient_id:
          type: string
        practitioner_id:
          type: string
        encounter_type:
          type: string
        reason_for_visit:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/CustomEncounterLocation'
        status:
          type: string
        diagnoses:
          type: array
          items:
            $ref: '#/components/schemas/CustomEncounterDiagnosis'
        services_provided:
          type: array
          items:
            type: string
        billing_info:
          $ref: '#/components/schemas/CustomEncounterBillingInfo'
        follow_up:
          $ref: '#/components/schemas/CustomEncounterFollowUp'
        note:
          type: string
    CustomEncounterLocation:
      type: object
      properties:
        facility_name:
          type: string
        room:
          type: string
        address:
          $ref: '#/components/schemas/CustomAddress'
    CustomEncounterDiagnosis:
      type: object
      properties:
        condition_id:
          type: string
        condition_name:
          type: string
    CustomEncounterBillingInfo:
      type: object
      properties:
        insurance_provider:
          type: string
        policy_number:
          type: string
        total_cost:
          type: number
        patient_responsibility:
          type: number
    CustomEncounterFollowUp:
      type: object
      properties:
        recommended:
          type: boolean
        next_appointment_date:
          type: string
          format: date-time
    CustomMedication:
      type: object
      properties:
        medication_id:
          type: string
        name:
          type: string
        brand_name:
          type: string
        form:
          type: string
        strength:
          type: string
        manufacturer:
          $ref: '#/components/schemas/CustomMedicationManufacturer'
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/CustomMedicationIngredient'
        package:
          $ref: '#/components/schemas/CustomMedicationPackage'
        status:
          type: string
        expiration_date:
          type: string
          format: date
        lot_number:
          type: string
    CustomMedicationManufacturer:
      type: object
      properties:
        name:
          type: string
        address:
          $ref: '#/components/schemas/CustomAddress'
    CustomMedicationIngredient:
      type: object
      properties:
        name:
          type: string
        strength:
          type: string
    CustomMedicationPackage:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
    CustomMedicationRequest:
      type: object
      properties:
        medication_request_id:
          type: number
        patient_id:
          type: string
        practitioner_id:
          type: string
        medication:
          $ref: '#/components/schemas/CustomMedicationRequestMedication'
        dosage_instruction:
          $ref: '#/components/schemas/CustomMedicationRequestDosage'
        dispense_request:
          $ref: '#/components/schemas/CustomMedicationRequestDispenseRequest'
        substitution_allowed:
          type: boolean
        reason_for_prescription:
          type: string
        status:
          type: string
        authored_on:
          type: string
          format: date
        note:
          type: string
    CustomMedicationRequestMedication:
      type: object
      properties:
        medication_id:
          type: string
        name:
          type: string
        strength:
          type: string
        form:
          type: string
...
    CustomPractitioner:
      type: object
      properties:
        practitioner_id:
          type: string
        name:
          $ref: '#/components/schemas/CustomPractitionerName'
        date_of_birth:
          type: string
          format: date
        gender:
          type: string
        contact_info:
          $ref: '#/components/schemas/CustomPractitionerContactInfo'
        license:
          $ref: '#/components/schemas/CustomPractitionerLicense'
        specialization:
          type: string
        years_of_experience:
          type: number
        languages_spoken:
          type: array
          items:
            type: string
        practice_location:
          $ref: '#/components/schemas/CustomPractitionerPracticeLocation'
        qualifications:
          type: array
          items:
            $ref: '#/components/schemas/CustomPractitionerQualification'
...
    CustomPractitionerRole:
      type: object
      properties:
        practitioner_role_id:
          type: string
        practitioner_id:
          type: string
        organization_id:
          type: string
        role:
          type: string
        specialty:
          type: string
        location:
          $ref: '#/components/schemas/CustomPractitionerRoleLocation'
        telecom:
          $ref: '#/components/schemas/CustomPractitionerRoleTelecom'
        available_hours:
          $ref: '#/components/schemas/CustomPractitionerRoleAvailableHours'
        service_period:
          $ref: '#/components/schemas/CustomPractitionerRoleServicePeriod'
...
    CustomOrganization:
      type: object
      properties:
        organization_id:
          type: string
        name:
          type: string
        type:
          type: string
        contact_info:
          $ref: '#/components/schemas/CustomOrganizationContactInfo'
        departments:
          type: array
          items:
            $ref: '#/components/schemas/CustomOrganizationDepartment'
        affiliations:
          type: array
          items:
            $ref: '#/components/schemas/CustomOrganizationAffiliation'
        services_offered:
          type: array
          items:
            type: string
        operating_hours:
          $ref: '#/components/schemas/CustomOrganizationOperatingHours'
...
    FhirBundle:
      type: object
      description: FHIR R5 Bundle (simplified)
      properties:
        resourceType:
          type: string
          example: Bundle
        type:
          type: string
        entry:
          type: array
          items:
            type: object
    CustomJsonPatientOnly:
      title: Patient
      allOf:
        - $ref: '#/components/schemas/CustomJsonInput'
        - type: object
          required: [patient]
    CustomJsonEncounterOnly:
      title: Encounter
      allOf:
        - $ref: '#/components/schemas/CustomJsonInput'
        - type: object
          required: [encounter]
    CustomJsonMedicationOnly:
      title: Medication
      allOf:
        - $ref: '#/components/schemas/CustomJsonInput'
        - type: object
          required: [medication]
    CustomJsonMedicationRequestOnly:
      title: MedicationRequest
      allOf:
        - $ref: '#/components/schemas/CustomJsonInput'
        - type: object
          required: [medication_request]
    CustomJsonPractitionerOnly:
      title: Practitioner
      allOf:
        - $ref: '#/components/schemas/CustomJsonInput'
        - type: object
          required: [practitioner]
    CustomJsonPractitionerRoleOnly:
      title: PractitionerRole
      allOf:
        - $ref: '#/components/schemas/CustomJsonInput'
        - type: object
          required: [practitioner_role]
    CustomJsonOrganizationOnly:
      title: Organization
      allOf:
        - $ref: '#/components/schemas/CustomJsonInput'
        - type: object
          required: [organization]
    CustomJsonFull:
      title: Full Bundle
      allOf:
        - $ref: '#/components/schemas/CustomJsonInput'
        - type: object
          required:
            - patient
            - encounter
            - medication
            - medication_request
            - practitioner
            - practitioner_role
            - organization